Add processDeeplink as util function
