+++ src/components/pages/InitializationRouter.tsx
@@ -1,10 +1,9 @@
 import * as React from "react";
 import { NavigationScreenProps, NavigationActions } from "react-navigation";
 import { Linking } from "react-native";
-import url from "url";
-
 import { Loading } from "../shared/Loading";
-import { RouteName, DEEPLINK_ROUTES } from "../shared/Navigation";
+import { RouteName } from "../shared/Navigation";
+import { processDeeplink } from "../shared/Deeplinking";
 
 type Props = { defaultRoute: RouteName } & NavigationScreenProps;
 
@@ -13,28 +12,13 @@ export class InitializationRouter extends React.Component<Props> {
     super(props);
   }
 
-  public componentDidMount() {
+  componentDidMount() {
     // Process deeplink -- we don't use react-navigation for this since it
     // doesn't support HTTPS links.
     Linking.getInitialURL()
       .then(link => {
         if (link) {
-          const deeplinkUrl = url.parse(link, true, true);
-          if (!deeplinkUrl.pathname) {
-            return;
-          }
-          const pathname = deeplinkUrl.pathname.replace(
-            "/",
-            ""
-          ) as keyof typeof DEEPLINK_ROUTES;
-          this.props.navigation.navigate(
-            "App",
-            {},
-            NavigationActions.navigate({
-              routeName: DEEPLINK_ROUTES[pathname],
-              params: deeplinkUrl.query
-            })
-          );
+          processDeeplink(link, this.props.navigation);
         } else {
           const defaultRoute = this.props.defaultRoute;
           if (defaultRoute) {
